refactor(avatar/user): 엔티티 경로 수정 #16
